requirements.txt
openai
python-dotenv
pydantic
httpx
tiktoken
